YES! HAHA! A whole 22 lines of code, better than nothing right?
+Now people can remove alliances and wars agenst   other guilds. They can't ally guilds from an opposing faction and they can war guilds from a friendly faction.

+Now, it's a little hard to test this by myself. Someone do it for me!

NOTE: Sorry but ATM I am not going to make wars like OSI right now. Ya know, where you kill so many of the enemy guild members you win thingy. I think I didn't do it in the first place because chrdeath.ecl is not run inside a package. If my lazy ass gets around to it I'll test to see if it does work, if it doesn't I'm going to bug the hell out of some peoples to make it happen! =D

I really just want to try and make this package a drop-n-compile so anyone can use it without a lot of hassle.

Alright, I'm going to stop typing now, this is really killing my script high and if I type anymore I'm not going to want to script for a few weeks. So, I'm going to leave it at that. Hopefully that was enough information, I slightly remember someone complaining about not enough info being in these little messages. So there you go, now my hand is cramping...